[Fix] PillData 수정
