--- SobokSobok/SobokSobok/Data/DTO/APIModel/Medicine/PillCount.swift
@@ -1,13 +1,0 @@
-//
-//  PillCount.swift
-//  SobokSobok
-//
-//  Created by 김승찬 on 2022/01/17.
-//
-
-import Foundation
-
-// MARK: - PillCount
-struct PillCount: Codable {
-    let pillCount: String
-}
