+++ SobokSobok/SobokSobok/Data/DTO/APIModel/Pill/PillCount.swift
@@ -1,0 +1,10 @@
+//
+//  PillCount.swift
+//  SobokSobok
+//
+//  Created by taekki on 2022/08/08.
+//
+
+struct PillCount: Codable {
+    let pillCount: Int
+}
